Extract common functions and use safe assignment functions of sup-dto

+++ src/test/c++/unit/unit_test_helper.h
@@ -64,8 +64,8 @@ public:
    */
   static const std::string Type;
 
-  static unsigned counter;
-  static unsigned GetCount();
+  static unsigned long counter;
+  static unsigned long GetCount();
 };
 
 class MockUI : public sup::sequencer::UserInterface
